+++ index.html
@@ -125,6 +125,6 @@
       </div>
     </section>
   </main>
-  <script src="./js/script.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
